Edited the battle sequences
It resets you if you lose…so good thing, you win again, but need to
find another way to save the pokemon stats, so it resets you but you
keep your pokemon skills…  uGH ugh flkjaweflkawef

tldr; fixed so you can play again if you lose, but it does so by
rerandomizing your stats.
so improve that so that the pokemon stays static.

+ add the padthai
+add the next steps

+leanify the code